Reorganize piano roll files
